+++ packages/obsidian/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 0.1.12 (2024-01-06)
+
+
+### Features
+
+* cache update ([59e98fa](https://github.com/rowthan/pagenote/commit/59e98fa15fa3b4826af6c4e389564b166cd049a9))
+
+
+
+
+
+## 0.1.11 (2023-12-31)
+
+**Note:** Version bump only for package @pagenote/obsidian
+
+
+
+
+
 ## 0.1.10 (2023-12-31)
 
 **Note:** Version bump only for package @pagenote/obsidian